Add support for "off" mode
Flair pucks do not support an "off" mode (that is stored in a separate on / off setting), but Home Assistant does support an "off" setting. Prior to this patch, using the off mode in Home Assistant did nothing (except return an error if you knew to look for it). With this patch, it now correctly maps the HA "off" mode to Flair's "off" power setting.